+++ apps/demo-app/CHANGELOG.md
@@ -1,5 +1,47 @@
 # demo-app
 
+## 1.0.0-beta.24
+
+### Patch Changes
+
+- Updated dependencies [6f7149f]
+  - sit-onyx@1.0.0-beta.24
+
+## 1.0.0-beta.23
+
+### Patch Changes
+
+- Updated dependencies [dae102e]
+  - sit-onyx@1.0.0-beta.23
+
+## 1.0.0-beta.22
+
+### Patch Changes
+
+- Updated dependencies [349f412]
+  - sit-onyx@1.0.0-beta.22
+
+## 1.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies [d6321d8]
+  - sit-onyx@1.0.0-beta.21
+
+## 1.0.0-beta.20
+
+### Patch Changes
+
+- Updated dependencies [4c73713]
+  - sit-onyx@1.0.0-beta.20
+
+## 1.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies [9570420]
+  - sit-onyx@1.0.0-beta.19
+
 ## 1.0.0-beta.18
 
 ### Patch Changes